refactor: extract prop-driven date range selection components
DateSelectionComboBox and a new DateRangePicker/DateRangeSelection pair
now take startDate/endDate/setDateRange props instead of reading the
global date store directly. GlobalDateRangeSelection,
GlobalDateRangePicker, and GlobalDateSelection become thin wrappers
that inject the global store hooks, so the public API and behavior are
unchanged while the presentational components can be bound to any date
store.
