+++ src/components/DateSelection/DateSelectionComboBox.tsx
@@ -2,7 +2,7 @@ import { useCallback, useId, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { getActivationDate } from '@utils/business'
-import { useGlobalDateRange, useGlobalDateRangeActions } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
+import { type DateRange } from '@providers/DateStoreProvider/internal/types'
 import { useLayerContext } from '@contexts/LayerContext/LayerContext'
 import { ComboBox } from '@ui/ComboBox/ComboBox'
 import { VStack } from '@ui/Stack/Stack'
@@ -14,13 +14,19 @@ type DateSelectionOption = {
   value: DatePreset
 }
 
-export const DateSelectionComboBox = ({ showLabel = false }: { showLabel?: boolean }) => {
+type DateSelectionComboBoxProps = {
+  startDate: Date
+  endDate: Date
+  setDateRange: (range: DateRange) => void
+  showLabel?: boolean
+}
+
+export const DateSelectionComboBox = ({ startDate, endDate, setDateRange, showLabel = false }: DateSelectionComboBoxProps) => {
   const { t } = useTranslation()
   const [lastPreset, setLastPreset] = useState<DatePreset | null>(null)
   const { business } = useLayerContext()
 
-  const dateRange = useGlobalDateRange({ dateSelectionMode: 'full' })
-  const { setDateRange } = useGlobalDateRangeActions()
+  const dateRange = useMemo(() => ({ startDate, endDate }), [startDate, endDate])
 
   const selectedPreset = presetForDateRange(dateRange, lastPreset, getActivationDate(business))
